Added diamond grid, updated welcome gallery
Experimented with CSS, figured out what I need to do in order to create a dynamically flowing diamond grid. I'll need to work on the nitty-gritty details, but I've got a good idea of the direction I want to go in.

Next steps include documenting the diamond-grid component, then actually implementing it.