More logic outlining
Outlining more of the logic behind rendering each row. I think the distinction between top/other rows (which impacts the use of a negative top-margin) might be unnecessary, if I just add a v-if statement.

I'll probably rework the algorithm for generating rows at some point, but let's not pre-optimize.